refactor: extract triathlon flow
